requirements.txt
google-generativeai == 0.8.2
streamlit==1.20.0
googletrans==4.0.0-rc1
geopy==2.3.0
requests==2.31.0
speechrecognition==3.8.1
pyttsx3==2.90
gTTS==2.3.2